SELC-2789 -modified the code to reduce complexity
